Fixed the print function in io module, which didn't include the print statement from the stdio file. Modified the max length of arrays and the way they are created

+++ include/types.h
@@ -54,7 +54,7 @@ typedef struct str{
 }Str;
 
 typedef struct array{
-    int16_t len;
+    int32_t len;
     Value *values;
 }Array;
